Refactoring DateConfig into an extension
