--- MobileWallet/Common/DateConfig.swift
@@ -1,91 +1,0 @@
-//  DateConfig.swift
-
-/*
-    Package MobileWallet
-    Created by Gugulethu on 2019/11/07
-    Using Swift 5.0
-    Running on macOS 10.15
-
-    Copyright 2019 The Tari Project
-
-    Redistribution and use in source and binary forms, with or
-    without modification, are permitted provided that the
-    following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright notice,
-    this list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above
-    copyright notice, this list of conditions and the following disclaimer in the
-    documentation and/or other materials provided with the distribution.
-
-    3. Neither the name of the copyright holder nor the names of
-    its contributors may be used to endorse or promote products
-    derived from this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
-    CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
-    INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
-    OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-    CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
-    NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
-    HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-    CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
-    OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
-import Foundation
-
-public class DateConfig {
-
-    /*
-     Creates a readable string from a given date
-     */
-    class func getRelativeDayValue(fromDate date: Date) -> String? {
-        if Calendar.current.isDateInToday(date) {
-            return NSLocalizedString("Today", comment: "Transaction list section heading")
-        } else if Calendar.current.isDateInYesterday(date) {
-            return NSLocalizedString("Yesterday", comment: "Transaction list section heading")
-        } else {
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = DateFormatter.dateFormat(fromTemplate: "MMM d, YYYY", options: 0, locale: NSLocale.current)
-            dateFormatter.timeZone = TimeZone.current
-
-            return dateFormatter.string(from: date)
-        }
-    }
-
-    /*
-     Creates a date and time string from a given date
-     */
-    class func getDateTimeString(date: Date) -> String? {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = DateFormatter.dateFormat(fromTemplate: "MMMM d YYYY h:mm a", options: 0, locale: NSLocale.current)
-        dateFormatter.timeZone = TimeZone.current
-
-        return dateFormatter.string(from: date)
-    }
-
-    /*
-     Checks if the given date falls in the current week
-     */
-    //Unused at the moment (tests need fixing)
-//    class func dateFallsInCurrentWeek(date: Date) -> Bool {
-//        let currentWeek = Calendar.current.component(Calendar.Component.weekOfYear, from: Date())
-//        let datesWeek = Calendar.current.component(Calendar.Component.weekOfYear, from: date)
-//        return currentWeek == datesWeek
-//    }
-
-    /*
-     Mocks dates in the past
-     */
-    class func mockPastDate(byDayValue value: Int) -> Date {
-        let today = Date.init()
-        return Calendar.current.date(byAdding: Calendar.Component.day, value: -value, to: today)!
-    }
-
-}
